refactor: 각 도메인별 entity 패키지 추가 후 파일 이동
