--- src/main/java/com/kakaotech/team18/backend_server/domain/application/Status.java
@@ -1,7 +1,0 @@
-package com.kakaotech.team18.backend_server.domain.application;
-
-public enum Status {
-    PENDING,
-    PASS,
-    FAIL,
-}
