+++ src/main/java/com/kakaotech/team18/backend_server/domain/application/entity/Application.java
@@ -1,7 +1,9 @@
-package com.kakaotech.team18.backend_server.domain.applicationFormField;
+package com.kakaotech.team18.backend_server.domain.application.entity;
 
 import com.kakaotech.team18.backend_server.domain.BaseEntity;
-import com.kakaotech.team18.backend_server.domain.applicationForm.ApplicationForm;
+import com.kakaotech.team18.backend_server.domain.applicationForm.entity.ApplicationForm;
+import com.kakaotech.team18.backend_server.domain.club.entity.Club;
+import com.kakaotech.team18.backend_server.domain.user.entity.Users;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -17,21 +19,27 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApplicationFormField extends BaseEntity {
+public class Application extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "application_form_id")
+    @Column(name = "application_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "application_form_id")
+    @JoinColumn(name = "user_id")
+    private Users users;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "club_id")
+    private Club club;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "applictaion_form_id")
     private ApplicationForm applicationForm;
 
-    private String question;
-
     @Enumerated(EnumType.STRING)
-    private FieldType fieldType;
+    private Status status;
 
-    private boolean isRequired;
+
 }
