+++ src/main/java/com/kakaotech/team18/backend_server/domain/application/entity/Status.java
@@ -1,0 +1,7 @@
+package com.kakaotech.team18.backend_server.domain.application.entity;
+
+public enum Status {
+    PENDING,
+    PASS,
+    FAIL,
+}
