+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationForm/entity/ApplicationForm.java
@@ -1,12 +1,10 @@
-package com.kakaotech.team18.backend_server.domain.clubMember;
+package com.kakaotech.team18.backend_server.domain.applicationForm.entity;
 
 import com.kakaotech.team18.backend_server.domain.BaseEntity;
-import com.kakaotech.team18.backend_server.domain.club.Club;
-import com.kakaotech.team18.backend_server.domain.user.Users;
+import com.kakaotech.team18.backend_server.domain.application.entity.Application;
+import com.kakaotech.team18.backend_server.domain.club.entity.Club;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,26 +16,22 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ClubMember extends BaseEntity {
+public class ApplicationForm extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "club_member_id")
+    @Column(name = "application_form_id")
     private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private Users user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
     private Club club;
 
-    @Enumerated(EnumType.STRING)
-    private Position position;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "application_id")
+    private Application application;
 
-    @Enumerated(EnumType.STRING)
-    private ActiveStatus activeStatus;
+    private String title;
 
-
+    private String description;
 }
