+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationFormField/entity/FieldType.java
@@ -1,0 +1,7 @@
+package com.kakaotech.team18.backend_server.domain.applicationFormField.entity;
+
+public enum FieldType {
+    TEXT,
+    RADIO,
+    CHECKBOX
+}
