+++ src/main/java/com/kakaotech/team18/backend_server/domain/clubMember/entity/ClubMember.java
@@ -1,10 +1,12 @@
-package com.kakaotech.team18.backend_server.domain.applicationForm;
+package com.kakaotech.team18.backend_server.domain.clubMember.entity;
 
 import com.kakaotech.team18.backend_server.domain.BaseEntity;
-import com.kakaotech.team18.backend_server.domain.application.Application;
-import com.kakaotech.team18.backend_server.domain.club.Club;
+import com.kakaotech.team18.backend_server.domain.club.entity.Club;
+import com.kakaotech.team18.backend_server.domain.user.entity.Users;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -16,22 +18,26 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApplicationForm extends BaseEntity {
+public class ClubMember extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "application_form_id")
+    @Column(name = "club_member_id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private Users user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
     private Club club;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "application_id")
-    private Application application;
+    @Enumerated(EnumType.STRING)
+    private Position position;
 
-    private String title;
+    @Enumerated(EnumType.STRING)
+    private ActiveStatus activeStatus;
 
-    private String description;
+
 }
