+++ src/main/java/com/kakaotech/team18/backend_server/domain/clubMember/entity/Position.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team18.backend_server.domain.clubMember;
+package com.kakaotech.team18.backend_server.domain.clubMember.entity;
 
 public enum Position {
     PRESIDENT,
